modify: '홈' deleted from NavBar, style folder refactoring in progress, detail header fixed to post.coverImages
